Adding player strafing
These changes add the ability for the player to strafe using the q and e
keys (assigned to a new "Strafe" axis).  There are animation issues that
will still need to be resolved.  We may also still need to tweak the
sensitivity, as it seems a little slower than moving forward or turning.